fix: prevent crash in determineDropResult (react-dnd#1281)
When target is undefined, determineDropResult crashes. I was not able to determine why `registry.getTarget` would return undefined. The issues happens sometimes when I am dragging and dropping quickly. Perhaps the drop target is getting unmounted as mentioned in #31.

I am happy to dig deeper if someone can point me in the right direction regarding which files might be relevant.

This fixes #1281.